split Disassemble
